// frontend/pages/api/Schema.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
  readonly "/subjects": {
    /** 被験者一覧取得 */
    readonly get: operations["index-subject"];
    /** 被験者作成 */
    readonly post: operations["create-subject"];
  };
  readonly "/subjects/{subject_id}": {
    /** 被験者詳細 */
    readonly get: operations["show-subject"];
    readonly parameters: {
      readonly path: {
        readonly subject_id: number;
      };
    };
  };
}

export interface definitions {
  readonly CreateSubjectRequest: {
    readonly name: string;
    readonly system_name: string;
  };
  readonly ListSubject: {
    readonly subjects?: readonly definitions["Subject"][];
  };
  readonly Subject: {
    readonly id: number;
    readonly name: string;
    readonly system_name: string;
  };
  readonly CreateSubjectResponse: {
    readonly id: number;
  };
}

export interface responses {}

export interface operations {
  /** 被験者一覧取得 */
  readonly "index-subject": {
    readonly parameters: {};
    readonly responses: {
      /** OK */
      readonly 200: {
        readonly schema: definitions["ListSubject"];
      };
    };
  };
  /** 被験者作成 */
  readonly "create-subject": {
    readonly parameters: {
      readonly body: {
        /** 被験者詳細 */
        readonly body: definitions["CreateSubjectRequest"];
      };
    };
    readonly responses: {
      /** Created */
      readonly 201: {
        readonly schema: definitions["CreateSubjectResponse"];
      };
    };
  };
  /** 被験者詳細 */
  readonly "show-subject": {
    readonly parameters: {
      readonly path: {
        /** subject id */
        readonly subject_id: number;
      };
    };
    readonly responses: {
      /** OK */
      readonly 200: {
        readonly schema: definitions["Subject"];
      };
    };
  };
}